requirements.txt
asgiref==3.5.2
Django==4.1.3
django-debug-toolbar==3.7.0
django-filter==22.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
